// src/database/types.ts
/**
 * This file was generated by kysely-codegen.
 * Please do not edit it manually.
 */

import type { ColumnType } from "kysely";

export type Generated<T> = T extends ColumnType<infer S, infer I, infer U>
  ? ColumnType<S, I | undefined, U>
  : ColumnType<T, T | undefined, T>;

export type Json = JsonValue;

export type JsonArray = JsonValue[];

export type JsonObject = {
  [x: string]: JsonValue | undefined;
};

export type JsonPrimitive = boolean | number | string | null;

export type JsonValue = JsonArray | JsonObject | JsonPrimitive;

export type Timestamp = ColumnType<Date, Date | string, Date | string>;

export interface EmailAttempts {
  created_at: Generated<Timestamp | null>;
  email_id: string;
  error: string | null;
  id: Generated<string>;
  provider: string;
  status: string;
}

export interface Emails {
  bcc: string[] | null;
  body: Json;
  cc: string[] | null;
  created_at: Generated<Timestamp | null>;
  error: string | null;
  from_address: string | null;
  id: string;
  reply_to: string | null;
  status: string;
  subject: string;
  to_address: string[];
  updated_at: Generated<Timestamp | null>;
}

export interface KnexMigrations {
  batch: number | null;
  id: Generated<number>;
  migration_time: Timestamp | null;
  name: string | null;
}

export interface KnexMigrationsLock {
  index: Generated<number>;
  is_locked: number | null;
}

export interface Providers {
  config: string;
  created_at: Generated<Timestamp | null>;
  id: Generated<string>;
  name: string;
  priority: number;
  status: Generated<string>;
  type: Generated<string>;
  updated_at: Generated<Timestamp | null>;
}

export interface DB {
  email_attempts: EmailAttempts;
  emails: Emails;
  knex_migrations: KnexMigrations;
  knex_migrations_lock: KnexMigrationsLock;
  providers: Providers;
}